requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-dotenv==1.0.1
redis==5.0.8
pytest==8.3.2
httpx==0.27.2
pydantic[email]==2.8.2